# apps/dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SimpleMailer legacy API",
    "version": "2.1.0",
    "description": "The current compatibility API. Server-to-server callers authenticate with x-api-key. Dashboard browsers use an HttpOnly session and a server-side proxy; the administrative API key is never exposed to JavaScript."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "LegacyApiKey": []
    }
  ],
...
  "paths": {
...
    "/api/send": {
      "post": {
        "tags": [
          "Mail"
        ],
        "summary": "Accept and enqueue one mail job",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "description": "Makes a mutation replay-safe. Reuse with materially different content returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job persisted and accepted for enqueue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, invalid request, or validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid legacy API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Referenced account or template does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key was used with materially different content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the route limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Process-local development rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until another request may be attempted",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence or service failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "RabbitMQ publication failed; the persisted job remains enqueue-pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Mail"
        ],
        "summary": "Find a job by idempotency/enqueue key",
        "parameters": [
          {
            "in": "query",
            "name": "enqueueKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, invalid request, or validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid legacy API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No job exists for the key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "List SMTP accounts",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid legacy API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create an SMTP account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, invalid request, or validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid legacy API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the route limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence or service failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "LegacyApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Temporary server-to-server administrative credential."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "success",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {},
          "error": {
            "type": "string",
            "description": "Temporary legacy alias for message"
          },
          "fields": {
            "description": "Temporary legacy alias for validation details"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MailJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accountId",
          "templateId",
          "recipient",
          "values"
        ],
        "properties": {
          "accountId": {
            "type": "string",
            "minLength": 1
          },
          "templateId": {
            "type": "string",
            "minLength": 1
          },
          "recipient": {
            "type": "string",
            "format": "email"
          },
          "values": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "SendAcceptedResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "jobId",
          "status"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENQUEUE_PENDING",
              "QUEUED",
              "PROCESSING",
              "PENDING",
              "SENT",
              "FAILED",
              "RETRYING",
              "DEAD",
              "DELIVERY_UNCERTAIN"
            ]
          }
        }
      },
      "SendStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "id",
          "status",
          "retryCount",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENQUEUE_PENDING",
              "QUEUED",
              "PROCESSING",
              "PENDING",
              "SENT",
              "FAILED",
              "RETRYING",
              "DEAD",
              "DELIVERY_UNCERTAIN"
            ]
          },
          "retryCount": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AccountRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "username",
          "password",
          "emailHost",
          "emailPort"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "writeOnly": true
          },
          "emailHost": {
            "type": "string",
            "minLength": 1
          },
          "emailPort": {
            "type": "integer",
            "minimum": 1,
            "maximum": 65535
          }
        }
      },
      "AccountSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "username",
          "emailHost",
          "emailPort",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "emailHost": {
            "type": "string"
          },
          "emailPort": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}